recherche d'entité par un de leur attribut + creation de tous les controlleurs necessaire

+++ src/main/java/com/example/springjee/entities/TypeProduit.java
@@ -14,7 +14,7 @@ public class TypeProduit {
     @Id
     @Column(name = "typeProduit_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "typeProduit_nom", nullable = false)
     private String nom;
@@ -45,9 +45,9 @@ public class TypeProduit {
     //////GETTER AND SETTER//////
     /////////////////////////////
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
